pyproject.toml
[project]
name = "task-execution-agent"
version = "0.1.0"
description = "Task Execution AI Agent"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.28.0",
    "openai>=2.36.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
